Added support to change the color of the ball
Added support to change the color of the ball via the color and
color_colorname broadcasts.  Also you can now toggle the back LED via
the back_led_on and back_led_off broadcasts.

Additionally retrieving the balls textual name via the bluetooth info
and adding that to the logging commands going to console (this is a
little buggy in that it doesn't always find the name).