Update RGB sensor values when using color_name
I was hoping to update Scratch's global variables via the RSC program,
but that doesn't seem to be available in the protocol.  I can update
'sensor values' though, so when you use the color_name logic, and when
it retrieves the user LED it will update an r, g, and b sensor value.

The Sphero doesn't not return it's current color for some reason so I
added a method to Sphero class and just retrieve it separately, not
elegant but accomplished what I wanted to do with the test?